add fee tests
